Replace lowrisc ibex with CHERI-enabled fork
It is not sufficient to clone the CHERI-enabled version into a new
directory because the fusesoc names will clash, so this replaces the old
vendored checkout.

This isn't the cleanest approach to replace the core and a better
approach would be to vendor it into its own folder and update all the
file/module/core names using a script, but this would require more
intrusive changes in opentitan and it's not yet possible to
automatically run a script using the vendor tool.